have a dedicated job for scheduling journal completed events
