+++ app/services/notifications/create_from_journal_service.rb
@@ -29,25 +29,234 @@
 #++
 
 class Notifications::CreateFromJournalService
+  MENTION_USER_ID_PATTERN =
+    '<mention[^>]*(?:data-type="user"[^>]*data-id="(\d+)")|(?:data-id="(\d+)"[^>]*data-type="user")[^>]*>)|(?:\buser#(\d+)\b'
+      .freeze
+  MENTION_USER_LOGIN_PATTERN =
+    '\buser:"(.+?)"'.freeze
+  MENTION_GROUP_ID_PATTERN =
+    '<mention[^>]*(?:data-type="group"[^>]*data-id="(\d+)")|(?:data-id="(\d+)"[^>]*data-type="group")[^>]*>)|(?:\bgroup#(\d+)\b'
+      .freeze
+  MENTION_PATTERN = Regexp.new("(?:#{MENTION_USER_ID_PATTERN})|(?:#{MENTION_USER_LOGIN_PATTERN})|(?:#{MENTION_GROUP_ID_PATTERN})")
+
   class << self
-    def call(journal, send_mails)
-      enqueue_completed_job(journal, send_mails) if supported?(journal)
+    def call(journal, send_notifications)
+      return nil if abort_sending?(journal, send_notifications)
+
+      notification_receivers(journal).each do |recipient_id, channel_reasons|
+        create_notification(journal,
+                            recipient_id,
+                            channel_reasons)
+      end
     end
 
     private
 
-    def enqueue_completed_job(journal, send_mails)
-      Notifications::JournalCompletedJob
-        .set(wait_until: delivery_time)
-        .perform_later(journal.id, send_mails)
+    def create_notification(journal, recipient_id, channel_reasons)
+      notification_attributes = {
+        recipient_id: recipient_id,
+        project: journal.project,
+        resource: journal.journable,
+        journal: journal,
+        actor: user_with_fallback(journal)
+      }.merge(channel_attributes(channel_reasons))
+
+      Notifications::CreateService
+        .new(user: user_with_fallback(journal))
+        .call(notification_attributes)
     end
 
-    def delivery_time
-      Setting.journal_aggregation_time_minutes.to_i.minutes.from_now
+    def channel_attributes(channel_reasons)
+      {
+        read_mail: channel_reasons.keys.include?('mail') ? false : nil,
+        read_mail_digest: channel_reasons.keys.include?('mail_digest') ? false : nil,
+        read_ian: channel_reasons.keys.include?('in_app') ? false : nil,
+        reason_ian: channel_reasons['in_app']&.first,
+        reason_mail: channel_reasons['mail']&.first,
+        reason_mail_digest: channel_reasons['mail_digest']&.first
+      }
     end
 
-    def supported?(journal)
-      %w(WorkPackage WikiContent).include?(journal.journable_type)
+    def notification_receivers(journal)
+      receivers = receivers_hash
+
+      %i(mentioned involved watched subscribed commented created processed prioritized scheduled).each do |reason|
+        add_receivers_by_reason(receivers, journal, reason)
+      end
+
+      remove_self_recipient(receivers, journal)
+
+      receivers
+    end
+
+    def add_receivers_by_reason(receivers, journal, reason)
+      add_receiver(receivers, send(:"settings_of_#{reason}", journal), reason)
+    end
+
+    def settings_of_mentioned(journal)
+      applicable_settings(mentioned_ids(journal),
+                          journal.data.project,
+                          :mentioned)
+    end
+
+    def settings_of_involved(journal)
+      scope = User
+                .where(id: group_or_user_ids(journal.data.assigned_to))
+                .or(User.where(id: group_or_user_ids(journal.data.responsible)))
+
+      applicable_settings(scope,
+                          journal.data.project,
+                          :involved)
+    end
+
+    def settings_of_subscribed(journal)
+      project = journal.data.project
+
+      applicable_settings(User.notified_on_all(project),
+                          project,
+                          :all)
+    end
+
+    def settings_of_watched(journal)
+      work_package = journal.journable
+
+      applicable_settings(work_package.watcher_users,
+                          work_package.project,
+                          :watched)
+    end
+
+    def settings_of_commented(journal)
+      return NotificationSetting.none unless journal.notes?
+
+      applicable_settings(User.all,
+                          journal.data.project,
+                          :work_package_commented)
+    end
+
+    def settings_of_created(journal)
+      return NotificationSetting.none unless journal.initial?
+
+      applicable_settings(User.all,
+                          journal.data.project,
+                          :work_package_created)
+    end
+
+    def settings_of_processed(journal)
+      return NotificationSetting.none unless !journal.initial? && journal.details.has_key?(:status_id)
+
+      applicable_settings(User.all,
+                          journal.data.project,
+                          :work_package_processed)
+    end
+
+    def settings_of_prioritized(journal)
+      return NotificationSetting.none unless !journal.initial? && journal.details.has_key?(:priority_id)
+
+      applicable_settings(User.all,
+                          journal.data.project,
+                          :work_package_prioritized)
+    end
+
+    def settings_of_scheduled(journal)
+      if journal.initial? || !(journal.details.has_key?(:start_date) || journal.details.has_key?(:due_date))
+        return NotificationSetting.none
+      end
+
+      applicable_settings(User.all,
+                          journal.data.project,
+                          :work_package_scheduled)
+    end
+
+    def applicable_settings(user_scope, project, reason)
+      NotificationSetting
+        .applicable(project)
+        .where(reason => true)
+        .where(user: user_scope.where(id: User.allowed(:view_work_packages, project)))
+    end
+
+    def text_for_mentions(journal)
+      potential_text = ""
+      potential_text << journal.notes if journal.try(:notes)
+
+      %i[description subject].each do |field|
+        details = journal.details[field]
+
+        if details.present?
+          potential_text << "\n#{Redmine::Helpers::Diff.new(*details.reverse).additions.join(' ')}"
+        end
+      end
+      potential_text
+    end
+
+    def mentioned_ids(journal)
+      matches = mention_matches(journal)
+
+      base_scope = User
+                     .includes(:groups)
+                     .references(:groups_users)
+
+      by_id = base_scope.where(id: matches[:user_ids])
+      by_login = base_scope.where(login: matches[:user_login_names])
+      by_group = base_scope.where(groups_users: { id: matches[:group_ids] })
+
+      by_id
+        .or(by_login)
+        .or(by_group)
+    end
+
+    def send_notification?(journal, send_notifications)
+      send_notifications && ::UserMailer.perform_deliveries
+    end
+
+    def mention_matches(journal)
+      text = text_for_mentions(journal)
+
+      user_ids_tag_after,
+        user_ids_tag_before,
+        user_ids_hash,
+        user_login_names,
+        group_ids_tag_after,
+        group_ids_tag_before,
+        group_ids_hash = text
+                           .scan(MENTION_PATTERN)
+                           .transpose
+                           .each(&:compact!)
+
+      {
+        user_ids: [user_ids_tag_after, user_ids_tag_before, user_ids_hash].flatten.compact,
+        user_login_names: [user_login_names].flatten.compact,
+        group_ids: [group_ids_tag_after, group_ids_tag_before, group_ids_hash].flatten.compact
+      }
+    end
+
+    def abort_sending?(journal, send_notifications)
+      !send_notification?(journal, send_notifications) || journal.noop?
+    end
+
+    def group_or_user_ids(association)
+      association.is_a?(Group) ? association.user_ids : association&.id
+    end
+
+    def user_with_fallback(journal)
+      journal.user || DeletedUser.first
+    end
+
+    def add_receiver(receivers, collection, reason)
+      collection.each do |notification|
+        receivers[notification.user_id][notification.channel] << reason
+      end
+    end
+
+    def remove_self_recipient(receivers, journal)
+      receivers.delete(journal.user_id) if receivers[journal.user_id] && !user_with_fallback(journal).pref.self_notified?
+    end
+
+    def receivers_hash
+      Hash.new do |hash, user|
+        hash[user] = Hash.new do |channel_hash, channel|
+          channel_hash[channel] = []
+        end
+      end
     end
   end
 end
