+++ spec/features/notifications/notification_center/notification_center_spec.rb
@@ -1,29 +1,29 @@
 require 'spec_helper'
 require 'support/components/notifications/center'
 
-describe "Notification center", type: :feature, js: true do
-  shared_let(:project) { FactoryBot.create :project }
-  shared_let(:work_package) { FactoryBot.create :work_package, project: project }
-  shared_let(:second_work_package) { FactoryBot.create :work_package, project: project }
-  shared_let(:recipient) do
+describe "Notification center", type: :feature, js: true, with_settings: { journal_aggregation_time_minutes: 0 } do
+  # Notice that the setup in this file here is not following the normal rules as
+  # it also tests notification creation.
+  let!(:project) { FactoryBot.create :project }
+  let!(:recipient) do
+    # Needs to take place before the work package is created so that the notification listener is set up
     FactoryBot.create :user,
                       member_in_project: project,
                       member_with_permissions: %i[view_work_packages]
   end
-  shared_let(:notification) do
-    FactoryBot.create :notification,
-                      recipient: recipient,
-                      project: project,
-                      resource: work_package,
-                      journal: work_package.journals.last
+  let!(:work_package) do
+    FactoryBot.create :work_package, project: project
   end
-
-  shared_let(:second_notification) do
-    FactoryBot.create :notification,
-                      recipient: recipient,
-                      project: project,
-                      resource: second_work_package,
-                      journal: work_package.journals.last
+  let!(:work_package2) do
+    FactoryBot.create :work_package, project: project
+  end
+  let!(:notification) do
+    # Will have been created via the JOURNAL_CREATED event listeners
+    work_package.journals.first.notifications.first
+  end
+  let!(:notification2) do
+    # Will have been created via the JOURNAL_CREATED event listeners
+    work_package2.journals.first.notifications.first
   end
 
   let(:center) { ::Components::Notifications::Center.new }
@@ -40,7 +40,7 @@ describe "Notification center", type: :feature, js: true do
       center.open
 
       center.expect_work_package_item notification
-      center.expect_work_package_item second_notification
+      center.expect_work_package_item notification2
       center.click_item notification
       split_screen.expect_open
 
@@ -57,7 +57,7 @@ describe "Notification center", type: :feature, js: true do
       center.open
 
       center.expect_work_package_item notification
-      center.expect_work_package_item second_notification
+      center.expect_work_package_item notification2
       center.mark_all_read
 
       center.expect_bell_count 0
@@ -65,7 +65,7 @@ describe "Notification center", type: :feature, js: true do
       expect(notification.read_ian).to be_truthy
 
       center.expect_no_item notification
-      center.expect_no_item second_notification
+      center.expect_no_item notification2
     end
 
     it 'can open the split screen of the notification' do
@@ -75,8 +75,9 @@ describe "Notification center", type: :feature, js: true do
 
       center.click_item notification
       split_screen.expect_open
+
       center.expect_item_not_read notification
-      center.expect_work_package_item second_notification
+      center.expect_work_package_item notification2
 
       center.mark_notification_as_read notification
 
@@ -90,26 +91,41 @@ describe "Notification center", type: :feature, js: true do
 
       center.open
       center.expect_no_item notification
-      center.expect_work_package_item second_notification
+      center.expect_work_package_item notification2
     end
 
     context 'with multiple notifications per work package' do
       # In this context we have four notifications for two work packages.
-      shared_let(:third_notification) do
-        FactoryBot.create :notification,
-                          recipient: recipient,
-                          project: project,
-                          resource: second_work_package,
-                          journal: work_package.journals.last
+      let(:notification3) do
+        work_package2.journal_notes = 'A new notification is created here on wp 2'
+        work_package2.save!
+
+        # Will have been created via the JOURNAL_CREATED event listeners
+        work_package2.journals.last.notifications.first
       end
-      shared_let(:fourth_notification) do
-        FactoryBot.create :notification,
-                          recipient: recipient,
-                          project: project,
-                          resource: work_package,
-                          journal: work_package.journals.last
+      let(:notification4) do
+        work_package.subject = 'Changing the subject'
+        work_package.save!
+
+        # Will have been created via the JOURNAL_CREATED event listeners
+        work_package.journals.last.notifications.first
       end
-      let(:second_split_screen) { ::Pages::SplitWorkPackage.new second_work_package }
+      let(:split_screen2) { ::Pages::SplitWorkPackage.new work_package2 }
+
+      before do
+        # The notifications need to be created as a different user
+        # as they are otherwise swallowed to avoid self notification.
+        allow(User)
+          .to receive(:current)
+                .and_return(work_package2.author)
+
+        notification3
+        notification4
+
+        allow(User)
+          .to receive(:current)
+                .and_return(recipient)
+      end
 
       it 'aggregates notifications per work package and sets all as read when opened' do
         visit home_path
@@ -119,28 +135,35 @@ describe "Notification center", type: :feature, js: true do
         center.expect_number_of_notifications 2
 
         # Click on first list item, which should be the youngest notification
-        center.click_item fourth_notification
+        center.click_item notification4
 
         split_screen.expect_open
-        center.mark_notification_as_read fourth_notification
+        center.mark_notification_as_read notification4
 
         retry_block do
-          notification.reload
-          raise "Expected notification to be marked read" unless notification.read_ian
+          notification4.reload
+          raise "Expected notification to be marked read" unless notification4.read_ian
         end
 
-        expect(second_notification.reload.read_ian).to be_falsey
-        expect(third_notification.reload.read_ian).to be_falsey
-        expect(fourth_notification.reload.read_ian).to be_truthy
+        expect(notification.reload.read_ian).to be_truthy
+        expect(notification2.reload.read_ian).to be_falsey
+        expect(notification3.reload.read_ian).to be_falsey
+        expect(notification4.reload.read_ian).to be_truthy
 
         # Click on second list item, which should be the youngest notification that does
         # not belong to the work package that represents the first list item.
-        center.click_item third_notification
+        center.click_item notification3
 
-        second_split_screen.expect_open
-        center.mark_notification_as_read third_notification
-        expect(second_notification.reload.read_ian).to be_truthy
-        expect(third_notification.reload.read_ian).to be_truthy
+        split_screen2.expect_open
+        center.mark_notification_as_read notification3
+
+        retry_block do
+          notification3.reload
+          raise "Expected notification to be marked read" unless notification3.read_ian
+        end
+
+        expect(notification2.reload.read_ian).to be_truthy
+        expect(notification3.reload.read_ian).to be_truthy
       end
     end
   end
