+++ spec/workers/notifications/journal_completed_job_integration_spec.rb
@@ -30,7 +30,7 @@
 
 require 'spec_helper'
 
-describe Notifications::JournalCompletedJob, type: :model do
+describe Notifications::JournalCompletedJob, 'integration', type: :model do
   let(:project) { FactoryBot.create(:project) }
   let(:permissions) { [:view_work_packages] }
   let(:recipient) do
